UT-Refactor2.Move FE UT case
