+++ fe/fe-core/src/test/java/com/starrocks/sql/plan/AggregateTest.java
@@ -2,5 +2,164 @@
 
 package com.starrocks.sql.plan;
 
+import com.starrocks.common.FeConstants;
+import org.junit.Assert;
+import org.junit.Test;
+
 public class AggregateTest extends PlanTestBase {
+    @Test
+    public void testHaving() throws Exception {
+        String sql = "select v2 from t0 group by v2 having v2 > 0";
+        String planFragment = getFragmentPlan(sql);
+        Assert.assertTrue(planFragment.contains("PREDICATES: 2: v2 > 0"));
+
+        sql = "select sum(v1) from t0 group by v2 having v2 > 0";
+        planFragment = getFragmentPlan(sql);
+        Assert.assertTrue(planFragment.contains("PREDICATES: 2: v2 > 0"));
+
+        sql = "select sum(v1) from t0 group by v2 having sum(v1) > 0";
+        planFragment = getFragmentPlan(sql);
+        Assert.assertTrue(planFragment.contains("having: 4: sum > 0"));
+    }
+
+    @Test
+    public void testCountDistinctBitmapHll() throws Exception {
+        String sql = "select count(distinct v1), count(distinct v2), count(distinct v3), count(distinct v4), " +
+                "count(distinct b1), count(distinct b2), count(distinct b3), count(distinct b4) from test_object;";
+        getFragmentPlan(sql);
+
+        sql = "select count(distinct v1), count(distinct v2), " +
+                "count(distinct h1), count(distinct h2) from test_object";
+        getFragmentPlan(sql);
+    }
+
+    @Test
+    public void testHaving2() throws Exception {
+        String sql = "SELECT 8 from t0 group by v1 having avg(v2) < 63;";
+        String planFragment = getFragmentPlan(sql);
+        Assert.assertTrue(planFragment.contains("having: 4: avg < 63.0"));
+    }
+
+    @Test
+    public void testGroupByNull() throws Exception {
+        String sql = "select count(*) from test_all_type group by null";
+        String planFragment = getFragmentPlan(sql);
+        Assert.assertTrue(planFragment.contains("<slot 11> : NULL"));
+    }
+
+    @Test
+    public void testSumDistinctConst() throws Exception {
+        String sql = "select sum(2), sum(distinct 2) from test_all_type";
+        String thriftPlan = getThriftPlan(sql);
+        Assert.assertTrue(thriftPlan.contains("function_name:multi_distinct_sum"));
+    }
+
+    @Test
+    public void testGroupByAsAnalyze() throws Exception {
+        String sql = "select BITOR(825279661, 1960775729) as a from test_all_type group by a";
+        String planFragment = getFragmentPlan(sql);
+        Assert.assertTrue(planFragment.contains("group by: 11: bitor"));
+    }
+
+    @Test
+    public void testHavingAsAnalyze() throws Exception {
+        String sql = "select count(*) as count1 from test_all_type having count1 > 1";
+        String planFragment = getFragmentPlan(sql);
+        Assert.assertTrue(planFragment.contains("having: 11: count > 1"));
+    }
+
+    @Test
+    public void testGroupByAsAnalyze2() throws Exception {
+        String sql = "select v1 as v2 from t0 group by v1, v2;";
+        String planFragment = getFragmentPlan(sql);
+        Assert.assertTrue(planFragment.contains("group by: 1: v1, 2: v2"));
+    }
+
+    @Test
+    public void testDistinctRedundant() throws Exception {
+        String sql = "SELECT DISTINCT + + v1, v1 AS col2 FROM t0;";
+        String planFragment = getFragmentPlan(sql);
+        Assert.assertTrue(planFragment.contains("  |  group by: 1: v1\n"));
+    }
+
+    @Test
+    public void testColocateAgg() throws Exception {
+        FeConstants.runningUnitTest = true;
+        String queryStr;
+        String explainString;
+        queryStr = "select k2, count(k3) from nocolocate3 group by k2";
+        explainString = getFragmentPlan(queryStr);
+        Assert.assertTrue(explainString.contains("  3:AGGREGATE (merge finalize)\n"
+                + "  |  output: count(4: count)\n"
+                + "  |  group by: 2: k2\n"
+                + "  |  \n"
+                + "  2:EXCHANGE\n"
+                + "\n"
+                + "PLAN FRAGMENT 2\n"
+                + " OUTPUT EXPRS:"));
+        FeConstants.runningUnitTest = false;
+    }
+
+    @Test
+    public void testDistinctWithGroupBy1() throws Exception {
+        connectContext.getSessionVariable().setNewPlanerAggStage(3);
+        String queryStr = "select avg(v1), count(distinct v1) from t0 group by v1";
+        String explainString = getFragmentPlan(queryStr);
+        Assert.assertTrue(explainString.contains(" 4:AGGREGATE (update finalize)\n" +
+                "  |  output: avg(4: avg), count(1: v1)\n" +
+                "  |  group by: 1: v1\n" +
+                "  |  \n" +
+                "  3:AGGREGATE (merge serialize)\n" +
+                "  |  output: avg(4: avg)\n" +
+                "  |  group by: 1: v1"));
+        connectContext.getSessionVariable().setNewPlanerAggStage(0);
+    }
+
+    @Test
+    public void testGroupBy2() throws Exception {
+        String queryStr = "select avg(v2) from t0 group by v2";
+        String explainString = getFragmentPlan(queryStr);
+        Assert.assertTrue(explainString.contains("  2:Project\n"
+                + "  |  <slot 4> : 4: avg\n"
+                + "  |  \n"
+                + "  1:AGGREGATE (update finalize)\n"
+                + "  |  output: avg(2: v2)\n"
+                + "  |  group by: 2: v2\n"
+                + "  |  \n"
+                + "  0:OlapScanNode"));
+    }
+
+
+    @Test
+    public void testAggregateConst() throws Exception {
+        String sql = "select 'a', v2, sum(v1) from t0 group by 'a', v2; ";
+        String plan = getFragmentPlan(sql);
+        Assert.assertTrue(plan.contains("  2:Project\n"
+                + "  |  <slot 2> : 2: v2\n"
+                + "  |  <slot 5> : 5: sum\n"
+                + "  |  <slot 6> : 'a'\n"
+                + "  |  \n"
+                + "  1:AGGREGATE (update finalize)\n"
+                + "  |  output: sum(1: v1)\n"
+                + "  |  group by: 2: v2\n"));
+    }
+
+    @Test
+    public void testAggregateAllConst() throws Exception {
+        String sql = "select 'a', 'b' from t0 group by 'a', 'b'; ";
+        String plan = getFragmentPlan(sql);
+        Assert.assertTrue(plan.contains("  3:Project\n"
+                + "  |  <slot 4> : 4: expr\n"
+                + "  |  <slot 6> : 'b'\n"
+                + "  |  \n"
+                + "  2:AGGREGATE (update finalize)\n"
+                + "  |  group by: 4: expr\n"
+                + "  |  \n"
+                + "  1:Project\n"
+                + "  |  <slot 4> : 'a'\n"
+                + "  |  \n"
+                + "  0:OlapScanNode\n"
+                + "     TABLE: t0"));
+    }
+
 }
