+++ fe/fe-core/src/test/java/com/starrocks/sql/plan/SubqueryTest.java
@@ -2,5 +2,39 @@
 
 package com.starrocks.sql.plan;
 
+import org.junit.Assert;
+import org.junit.Test;
+
 public class SubqueryTest extends PlanTestBase {
+
+    @Test
+    public void testCountConstantWithSubquery() throws Exception {
+        String sql = "SELECT 1 FROM (SELECT COUNT(1) FROM t0 WHERE false) t;";
+        String thriftPlan = getThriftPlan(sql);
+        Assert.assertTrue(thriftPlan.contains("function_name:count"));
+    }
+
+    @Test
+    public void testSubqueryGatherJoin() throws Exception {
+        String sql = "select t1.v5 from (select * from t0 limit 1) as x inner join t1 on x.v1 = t1.v4";
+        String plan = getFragmentPlan(sql);
+        Assert.assertTrue(plan.contains(" OUTPUT EXPRS:\n"
+                + "  PARTITION: RANDOM\n"
+                + "\n"
+                + "  STREAM DATA SINK\n"
+                + "    EXCHANGE ID: 02\n"
+                + "    UNPARTITIONED\n"
+                + "\n"
+                + "  1:OlapScanNode\n"
+                + "     TABLE: t0"));
+    }
+
+    @Test
+    public void testSubqueryBroadJoin() throws Exception {
+        String sql = "select t1.v5 from t0 inner join[broadcast] t1 on cast(t0.v1 as int) = cast(t1.v4 as int)";
+        String plan = getFragmentPlan(sql);
+        Assert.assertTrue(plan.contains("  |  equal join conjunct: 7: cast = 8: cast\n"));
+        Assert.assertTrue(plan.contains("<slot 7> : CAST(1: v1 AS INT)"));
+        Assert.assertTrue(plan.contains("<slot 8> : CAST(4: v4 AS INT)"));
+    }
 }
