+++ fe/fe-core/src/test/java/com/starrocks/sql/plan/WindowTest.java
@@ -2,5 +2,79 @@
 
 package com.starrocks.sql.plan;
 
+import com.starrocks.sql.analyzer.SemanticException;
+import org.junit.Assert;
+import org.junit.Test;
+
 public class WindowTest extends PlanTestBase {
+    @Test
+    public void testLagWindowFunction() throws Exception {
+        String sql = "select lag(id_datetime, 1, '2020-01-01') over(partition by t1c) from test_all_type;";
+        String plan = getThriftPlan(sql);
+        Assert.assertTrue(plan.contains("signature:lag(DATETIME, BIGINT, DATETIME)"));
+
+        sql = "select lag(id_decimal, 1, 10000) over(partition by t1c) from test_all_type;";
+        plan = getThriftPlan(sql);
+        String expectSlice = "fn:TFunction(name:TFunctionName(function_name:lag), binary_type:BUILTIN," +
+                " arg_types:[TTypeDesc(types:[TTypeNode(type:SCALAR, scalar_type:TScalarType(type:DECIMAL64," +
+                " precision:10, scale:2))])], ret_type:TTypeDesc(types:[TTypeNode(type:SCALAR, " +
+                "scalar_type:TScalarType(type:DECIMAL64, precision:10, scale:2))]), has_var_args:false, " +
+                "signature:lag(DECIMAL64(10,2))";
+        Assert.assertTrue(plan.contains(expectSlice));
+
+        sql = "select lag(null, 1,1) OVER () from t0";
+        plan = getFragmentPlan(sql);
+        Assert.assertTrue(plan.contains("functions: [, lag(NULL, 1, 1), ]"));
+
+        sql = "select lag(id_datetime, 1, '2020-01-01xxx') over(partition by t1c) from test_all_type;";
+        expectedEx.expect(SemanticException.class);
+        expectedEx.expectMessage("The third parameter of `lag` can't not convert to DATETIME");
+        getThriftPlan(sql);
+    }
+
+    @Test
+    public void testPruneWindowColumn() throws Exception {
+        String sql = "select sum(t1c) from (select t1c, lag(id_datetime, 1, '2020-01-01') over( partition by t1c)" +
+                "from test_all_type) a ;";
+        String plan = getFragmentPlan(sql);
+        Assert.assertFalse(plan.contains("ANALYTIC"));
+    }
+
+    @Test
+    public void testWindowFunctionTest() throws Exception {
+        String sql = "select sum(id_decimal - ifnull(id_decimal, 0)) over (partition by t1c) from test_all_type";
+        String plan = getThriftPlan(sql);
+        Assert.assertTrue(
+                plan.contains("decimal_literal:TDecimalLiteral(value:0, integer_value:00 00 00 00 00 00 00 00)"));
+    }
+
+    @Test
+    public void testSameWindowFunctionReuse() throws Exception {
+        String sql = "select sum(v1) over() as c1, sum(v1) over() as c2 from t0";
+        String plan = getFragmentPlan(sql);
+        Assert.assertTrue(plan.contains("  3:Project\n" +
+                "  |  <slot 4> : 4: sum(1: v1)\n" +
+                "  |  \n" +
+                "  2:ANALYTIC\n" +
+                "  |  functions: [, sum(1: v1), ]"));
+
+        sql = "select c1+1, c2+2 from (select sum(v1) over() as c1, sum(v1) over() as c2 from t0) t";
+        plan = getFragmentPlan(sql);
+        Assert.assertTrue(plan.contains("  3:Project\n" +
+                "  |  <slot 5> : 4: sum(1: v1) + 1\n" +
+                "  |  <slot 6> : 4: sum(1: v1) + 2\n" +
+                "  |  \n" +
+                "  2:ANALYTIC\n" +
+                "  |  functions: [, sum(1: v1), ]"));
+
+        sql = "select c1+1, c2+2 from (select sum(v1) over() as c1, sum(v3) over() as c2 from t0) t";
+        plan = getFragmentPlan(sql);
+        Assert.assertTrue(plan.contains("  3:Project\n" +
+                "  |  <slot 6> : 4: sum(1: v1) + 1\n" +
+                "  |  <slot 7> : 5: sum(3: v3) + 2\n" +
+                "  |  \n" +
+                "  2:ANALYTIC\n" +
+                "  |  functions: [, sum(1: v1), ], [, sum(3: v3), ]"));
+    }
+
 }
